KPads c.d and tests rearragement

+++ Components/SignalMappers/MappedAddressBuilder.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 namespace Components.SignalMappers {
     public class MappedAddressBuilder {
 
-        private readonly HashSet<int> mappedAddresses = new HashSet<int>();
+        private readonly HashSet<ushort> mappedAddresses = new HashSet<ushort>();
         public MappedAddressBuilder Add(BitArray address) => Add(BitArrayHelper.ToUShortLE(address));
         public MappedAddressBuilder AddRange(BitArray addressFrom, BitArray addressTo) => AddRange(BitArrayHelper.ToUShortLE(addressFrom), BitArrayHelper.ToUShortLE(addressTo));
         public MappedAddressBuilder Remove(BitArray address) => Remove(BitArrayHelper.ToUShortLE(address));
@@ -17,7 +17,7 @@ namespace Components.SignalMappers {
         }
 
         public MappedAddressBuilder AddRange(ushort addressFrom, ushort addressTo) {
-            for (int i = addressFrom; i <= addressTo; i++) {
+            for (ushort i = addressFrom; i <= addressTo; i++) {
                 mappedAddresses.Add(i);
             }
 
@@ -30,14 +30,14 @@ namespace Components.SignalMappers {
         }
 
         public MappedAddressBuilder RemoveRange(ushort addressFrom, ushort addressTo) {
-            for (int i = addressFrom; i <= addressTo; i++) {
+            for (ushort i = addressFrom; i <= addressTo; i++) {
                 mappedAddresses.Remove(i);
             }
 
             return this;
         }
 
-        public HashSet<int> Build() {
+        public HashSet<ushort> Build() {
             return mappedAddresses;
         }
     }
